Fix capacitor power relays
The shield boost penalty was not being properly stacking penalized. This affects primarily chimeras, which are willing to accept the penalty to shield boost amount, since it improves their remote reps.

Near as I can tell, all items with this effect SHOULD be penalized. If the change is too broad, we can revert it.