Corrected the missile range formula
I was looking back at this for Tengu optimization in Incursion sites, and the results pyfa was giving me were not accurate to my observations. I looked back at the source code and the thread we derived the formula from, and we pulled one of their prototype formulas. The correct one occurred several posts later.

I've corrected the inaccuracy and found supplemental evidence to confirm the new formula. It's also much cleaner and easier to read now.